attempts to reduce memory usage

+++ teddybox/main/playback.h
@@ -50,5 +50,5 @@ int32_t pb_get_chapter(void);
 esp_err_t pb_play(const char *uri);
 esp_err_t pb_play_default_lang(uint32_t lang, uint32_t id);
 esp_err_t pb_play_default(uint32_t id);
-esp_err_t pb_play_content(uint32_t id);
+esp_err_t pb_play_content(uint64_t id);
 esp_err_t pb_stop();
